\name{dbt.Plot-package}
\alias{dbt.Plot-package}
\alias{dbt.Plot}
\docType{package}
\title{
  Plotting routines package
}
\description{
  Package contains different plotting functions like:
  - dbt_qqplot (Quantile/Quantile Plot)
  - qqcauchy: QQ-Plot compared to Cauchy distribution
  - qqchi2: QQ-Plot compared to chi square distribution
  - dbt_qqnorm: QQ-Plot compared to normal distribution
  - QQplotGMM: QQ-Plot compared to Gauss Mixture Model
  - qqpoisson: QQ-Plot compared to Poisson distribution
  - *fit: Plot with best-fit-line ('*' is wildcard)
}
\details{
\tabular{ll}{
Package: \tab dbt.Plot\cr
Type: \tab Package\cr
Version: \tab 1.0.9\cr
Date: \tab 2009-05-23\cr
LazyLoad: \tab yes\cr
}
}
\author{
Lyuben Manolov

}

\keyword{ package }