\name{dbt_qqnorm}
\alias{dbt_qqnorm}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ QQNorm }
\description{
  Quantile/Quantile-Plot compared to normal distribution
}
\usage{
dbt_qqnorm(x, Name, pstyle)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ Variable to be drawn }
  \item{Name}{ Description of the x-axis }
  \item{pstyle}{ Line type, see nplot }
}
\value{
  \item{qqx}{x-points for the qq-plot}
  \item{qqy}{y-points for the qq-plot}   
}
\author{databionics UnivMarburg}

