\name{dbt_screeplot}
\alias{dbt_screeplot}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ dbt_screeplot }
\description{
  Plots single and cumulative percentages of a vector
}
\usage{
dbt_screeplot(Data, Legend)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{Data}{ Data to be plotted }
  \item{Legend}{ Logical, whether to display or not the legend. TRUE by default. }
}
\value{no return value, creates plot in current graphic device}
\author{databionics UnivMarburg}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
%\keyword{ screeplot }
