\name{ecdfplot}
\alias{ecdfplot}
\title{ graphical parameters}
\usage{ecdfplot(Data,xlab = NULL, ylab = NULL, main = NULL,\ldots)}
\arguments{
\item{Data}{Data to create ecdf from}
\item{xlab}{Label of x axis}
\item{ylab}{Label of y axis}
\item{main}{Title of the Plot}
\item{\ldots}{Additional Parameters given to \code{\link{plot}}}
}
\description{Function to generate the plot of a ECDF of the given data.}
