% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{frame}
\alias{frame}
\title{See how a piece will look when framed}
\usage{
frame(
  df,
  paper_size = "A4",
  frame_size = "11x14",
  frame_depth = 0.5,
  matt_size = NULL,
  matt_color = "snow1",
  frame_color = "#B39671",
  show_paper_edges = FALSE
)
}
\arguments{
\item{df}{A data frame of coordinates (and optionally grouping and color
columns)}

\item{paper_size}{String specifying the size of paper used for plotting,
either in inches (e.g. "5x7") or one of "letter", "A4", "A3".}

\item{frame_size}{String specifying the size of the frame in inches (e.g.
"11x14")}

\item{frame_depth}{The thickness of the frame, in inches}

\item{matt_size}{Optional string specifying the size of size of the frame in
inches (e.g. "8x10")}

\item{matt_color}{Color of the matt}

\item{frame_color}{Color of the frame}

\item{show_paper_edges}{Logical; show red overlay where the edges of the
paper are}
}
\value{
A ggplot object
}
\description{
See how a piece will look when framed
}
