\name{hline}
\alias{hline}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Horizontal line }
\description{
  Draws horizontal line(s) to the current grapics device
}
\usage{
hline(y, x, c, w, lName)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{y}{ y-coordinate(s) of lines }
  \item{x}{ Optional: 2-element vector for x coordinates, [] or not given => x axis }
  \item{c}{ Optional: Colour, default is magenta }
  \item{w}{ Optional: Width of the line(s) }
  \item{lName}{ Optional: text written at right end of line }
}
\author{databionics UnivMarburg}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
%\keyword{ horizontal }
%\keyword{ line }