% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheCorDist.R
\name{pheCorDist}
\alias{pheCorDist}
\title{Create a figure showing distribution and correlation}
\usage{
pheCorDist(data, binwidth = NULL)
}
\arguments{
\item{data}{A data frame contain multiple phenotype of several samples}

\item{binwidth}{The width of the bins, same as geom_histogram}
}
\value{
A object of a figure
}
\description{
Create a figure showing distribution and correlation
}
\examples{
library(plot)
pheCorDist(sn)
}
