\name{plotacor}
\alias{plotacor}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Plot autocorrelation }
\description{
  Plots autocorrelation of a time series
}
\usage{
plotacor(Data, maxlag)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{Data}{ Data vector }
  \item{maxlag}{ Maximum lag }
}
\author{databionics UnivMarburg}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
