\name{PlotCircle}
\alias{PlotCircle}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ PlotCircle }
\description{
  Draws a circle in the current plot
}
\usage{
PlotCircle(Center, Radius, PlotSymbol)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{Center}{ Center= [Mx,My] }
  \item{Radius}{ Radius of the circle to be plotted }
  \item{PlotSymbol}{ Optional: see plot 'r-' by default }
}
\author{databionics UnivMarburg}


%\keyword{ plot }
%\keyword{ draw }
%\keyword{ circle }
