% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcode.R
\name{plotting_time}
\alias{plotting_time}
\title{Determine approximate plotting time from a data.frame}
\usage{
plotting_time(df, plotter_speed = 800)
}
\arguments{
\item{df}{A data.frame or tibble containing x and y coordinates and
optionally a group column}

\item{plotter_speed}{The drawing speed of the pen plotter, in millimeters per second}
}
\value{
The approximate plotting time (HH:MM:SS)
}
\description{
Determine approximate plotting time from a data.frame
}
