\name{plotyline}
\alias{plotyline}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Plot y-line }
\description{
  Draws horizontal lines in the current plot
}
\usage{
plotyline(Xpoints, PlotSymbol)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{Xpoints}{ x-values of the lines to be drawn }
  \item{PlotSymbol}{ Option: PlotSymbol  see plot 'r-' by default }
}
\details{   }
\value{   }
\references{  }
\author{databionics UnivMarburg}
\note{    }
\seealso{  }
\examples{ }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ plot }
\keyword{ horizontal }
\keyword{ line }
\keyword{ lines }
