\name{PlotZPixMatrix}
\alias{PlotZPixMatrix}
\title{ PlotPixMatrix(Data,Names,LowLim,HiLim); plot Best Matches as a pixel coulour picture }
\arguments{
\item{LineCoords[1:n,1]}{from Data(1:n,2) data cases in rows, variables in columns}
\item{ColCoords[1:n,2]}{from Data(1:n,2) data cases in rows, variables in columns}
\item{Z[1:n,1]}{Vektor}
\item{Tiling}{gekachelt=TRUE (default)}
\item{LowLim,HiLim}{limits for the color axis, Default: min und max}
}
\author{Reimplemented in R by MT, matlab: Version: ALU September 2007}
\description{
 PlotPixMatrix(Data,Names,LowLim,HiLim); plot Best Matches as a pixel coulour picture }
