\name{qqcauchy}
\alias{qqcauchy}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ QQCauchy }
\description{
  Quantile/Quantile-Plot compared to Cauchy distribution
}
\usage{
qqcauchy(x, Name, percentil, pstyle, filename)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ Variable to be drawn }
  \item{Name}{ Optional: descripton for the x-axis }
  \item{percentil}{ Percentage of data to be ignored as it's too extreme }
  \item{pstyle}{ Line type, for details see rplot }
  \item{filename}{ Name for the .eps file in which the plot to be saved }
}
\details{
  
}
\value{   }
\references{  }
\author{databionics UnivMarburg}
\note{   }
\seealso{  }
\examples{  }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
