\name{qqchi2}
\alias{qqchi2}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ QQ chi square }
\description{
  Quantile/Quantile-Plot compared to chi square distribution with df latitude grades
}
\usage{
qqchi2(x, Df, Name, pstyle, filename)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ Variable to be drawn }
  \item{Df}{ Latitude grades of the 'Chi square' distribution; 2 if not declared }
  \item{Name}{ Description of the x-axis; 'data' if not declared }
  \item{pstyle}{ Line type, see qqplot }
  \item{filename}{ Filename for the .eps file in which the plot can be saved }
}
\details{    }
\value{    }
\references{  }
\author{databionics UnivMarburg}
\note{   }
\seealso{  }
\examples{ }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
