\name{qqpoisson}
\alias{qqpoisson}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ QQPoisson }
\description{
  Quantile/Quantile-Plot compared to Poisson distribution
}
\usage{
qqpoisson(x, Lambda, Name, pstyle)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ Variable to be drawn }
  \item{Lambda}{ Lambda parameter of the Poisson distribution }
  \item{Name}{ Description of the x-axis; 'data' will be used if nothing declared }
  \item{pstyle}{ Line type; 'b.' if not declared }
}
\details{   }
\value{   }
\references{  }
\author{databionics UnivMarburg}
\note{    }
\seealso{ \code{\link{qqplot}} }
\examples{ }

