\name{qqpoissonfit}
\alias{qqpoissonfit}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ QQPoissonFit }
\description{
  Quantile/Quantile-Plot compared to Poisson distribution with best-fit-line
}
\usage{
qqpoissonfit(x, ux, ox, PoissonLambda, Name, pstyle)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ Variable to be drawn }
  \item{ux}{ Lower limit of the interval [ux, ox], in which a line will be interpolated }
  \item{ox}{ Upper limit of the interval [ux, ox], in which a line will be interpolated }
  \item{PoissonLambda}{ Lambda parameter of the Poisson distribution }
  \item{Name}{ Description of the x-axis; 'data' if not declared }
  \item{pstyle}{ Line type; 'b-' if not declared }
}
\details{   }
\value{    }
\references{  }
\author{databionics UnivMarburg}
\note{    }
\seealso{ \code{\link{qqnormfit}} }
\examples{ }
