\name{qquniform}
\alias{qquniform}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ QQUniform }
\description{
  Quantile/Quantile-Plot w.r.t. uniform / rectangular distribution
}
\usage{
qquniform(x, Name, pstyle, filename)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ Variable to be drawn }
  \item{Name}{ Description of the x-axis; 'data' if not declared }
  \item{pstyle}{ Plot symbol; '.' if not declared }
  \item{filename}{ Filename for the .eps file in which the plot can be saved }
}
\details{  }
\value{  }
\references{  }
\author{databionics UnivMarburg}
\note{  }
\seealso{ \code{\link{qqplot}} }
\examples{ }

\keyword{ plot }
\keyword{ qqplot }
\keyword{ qq-plot }
\keyword{ quantile/quantile-plot }
\keyword{ uniform }
\keyword{ distribution } 