% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcode.R
\name{read_gcode}
\alias{read_gcode}
\title{Read a g-code file into a data.frame}
\usage{
read_gcode(file)
}
\arguments{
\item{file}{A path to a g-code file}
}
\value{
A data.frame with columns for x, y, and grouping variable.
}
\description{
Read a g-code file into a data.frame
}
