% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcode.R
\name{resize_for_plotter}
\alias{resize_for_plotter}
\title{Title}
\usage{
resize_for_plotter(df, paper_width = 210, paper_height = 297, max_dim = NULL)
}
\arguments{
\item{max_dim}{}
}
\value{

}
\description{
Title
}
