\name{scatterpde}
\alias{scatterpde}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Scatter PDE }
\description{
  Scatterplot with PDE instead of histograms
}
\usage{
scatterpde(Data, Names, cls, ticks)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{Data}{ Data matrix }
  \item{Names}{ Optional: column names }
  \item{cls}{ Optional: class numbers }
  \item{ticks}{ Optional: logical value, whether to display ticks }
}
\details{  }
\value{   }
\references{  }
\author{databionics UnivMarburg}
\note{   }
\seealso{  }
\examples{ }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.

