\name{scatterplot}
\alias{scatterplot}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Scatterplot }
\description{
  Scatterplot of all columns in X, histograms on diagonal
}
\usage{
scatterplot(Data, Names, cls, ticks)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{Data}{ Data matrix }
  \item{Names}{ Optional: column names }
  \item{cls}{ Optional: class vector }
  \item{ticks}{ Optional: logical value whether to display ticks }
}
\details{  }
\value{  }
\references{  }
\author{databionics UnivMarburg}
\note{    }
\seealso{  }
\examples{ }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ scatterplot }
\keyword{ histogram }
