\name{shepard}
\alias{shepard}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Shepard diagram }
\description{  Draws the Sheppard diagram    }
\usage{
shepard(Data, Projections, Metric)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{Data}{ Data matrix - observations in lines, variables in rows  }
  \item{Projections}{ Projected data }
  \item{Metric}{ Optional: the metric used in the data space. This must be one of 'euclidean', 'maximum', 'manhattan', 'canberra', 'binary' or 'minkowsky', as specified in dist(stats).
                  By default Metric == 'sqeuclid' which results to a squared euclidian distance as used in the classical MDS publications }
}
\details{   }
\value{   }
\references{  }
\author{databionics UnivMarburg}
\note{   }
\seealso{  }
\examples{ }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
%\keyword{ sheppard }
