\name{tileplot}
\alias{tileplot}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Multiple plots in one window }
\description{
  Splits the current window into numOfCols columns and numOfRows rows plots
}
\usage{
tileplot(numOfRows, numOfCols)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{numOfRows}{ Number of rows }
  \item{numOfCols}{ Number of columns }
}
\details{
  
}
\value{
}
\references{  }
\author{databionics UnivMarburg}
\note{ }
\seealso{  }
\examples{  }
