\name{vline}
\alias{vline}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Vertical line }
\description{
  Plots vertical line(s)
}
\usage{
vline(x, y, c, w, lName)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ x coordinate(s) of lines }
  \item{y}{ Optional: 2-element vector for y coordinates, default or [] => y axis }
  \item{c}{ Colour }
  \item{w}{ Line width }
  \item{lName}{ Optional: Text, displayed at top of line }
}
\details{   }
\value{    }
\references{  }
\author{databionics UnivMarburg}
\note{    }
\seealso{  }
\examples{  }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ plot }
\keyword{ vertical }
\keyword{ line }
