% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{frame}
\alias{frame}
\title{Title}
\usage{
frame(
  df,
  paper_size = "A4",
  frame_size = "11x14",
  frame_depth = 0.5,
  matt_size = NULL,
  matt_color = "snow1",
  frame_color = "#B39671",
  show_paper_edges = FALSE
)
}
\arguments{
\item{show_paper_edges}{}
}
\value{

}
\description{
Title
}
